add: 모든 데이터 지우기 화면 구성
